+++ SourceCode/Data/Property.cs
@@ -9,13 +9,10 @@ namespace ModulesRegistry.Data
     {
         public Property()
         {
-            ModuleGables = new HashSet<ModuleGable>();
         }
 
         public int Id { get; set; }
         public string Name { get; set; }
         public string Value { get; set; }
-
-        public virtual ICollection<ModuleGable> ModuleGables { get; set; }
     }
 }